+++ src/java/ControllerAdmin/DeleteUser.java
@@ -2,23 +2,21 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
-package Controller;
+package ControllerAdmin;
 
 import Dal.DAOUser;
-import Model.User;
 import java.io.IOException;
 import java.io.PrintWriter;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.util.ArrayList;
 
 /**
  *
  * @author PiPi
  */
-public class UserList extends HttpServlet {
+public class DeleteUser extends HttpServlet {
 
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
@@ -31,10 +29,8 @@ public class UserList extends HttpServlet {
      */
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-//        DAOUser daoUser = new DAOUser();
-//        ArrayList<User> list = daoUser.getAllUsers();
-//        request.setAttribute("listU", list);
-//        request.getRequestDispatcher("/user").forward(request, response);
+        response.setContentType("text/html;charset=UTF-8");
+
     }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
@@ -49,11 +45,10 @@ public class UserList extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-//        processRequest(request, response);
-        DAOUser daoUser = new DAOUser();
-        ArrayList<User> list = daoUser.getAllUsers();
-        request.setAttribute("listU", list);
-        request.getRequestDispatcher("/BlogShop/admin/user/index.jsp").forward(request, response);
+        DAOUser dao = new DAOUser();
+        String id = request.getParameter("id");
+        dao.deleteUser(id);
+        response.sendRedirect(request.getContextPath() + "/user-list");
     }
 
     /**
