+++ src/java/ControllerAdmin/UserList.java
@@ -2,21 +2,23 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
-package Controller;
+package ControllerAdmin;
 
 import Dal.DAOUser;
+import Model.User;
 import java.io.IOException;
 import java.io.PrintWriter;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import java.util.ArrayList;
 
 /**
  *
  * @author PiPi
  */
-public class DeleteUser extends HttpServlet {
+public class UserList extends HttpServlet {
 
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
@@ -29,8 +31,10 @@ public class DeleteUser extends HttpServlet {
      */
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
-
+//        DAOUser daoUser = new DAOUser();
+//        ArrayList<User> list = daoUser.getAllUsers();
+//        request.setAttribute("listU", list);
+//        request.getRequestDispatcher("/user").forward(request, response);
     }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
@@ -45,10 +49,11 @@ public class DeleteUser extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        DAOUser dao = new DAOUser();
-        String id = request.getParameter("id");
-        dao.deleteUser(id);
-        response.sendRedirect(request.getContextPath() + "/user-list");
+//        processRequest(request, response);
+        DAOUser daoUser = new DAOUser();
+        ArrayList<User> list = daoUser.getAllUsers();
+        request.setAttribute("listU", list);
+        request.getRequestDispatcher("/BlogShop/admin/user/index.jsp").forward(request, response);
     }
 
     /**
